refactor(table): move table API to types module
