+++ ibis/expr/api.py
@@ -7,13 +7,12 @@ import datetime
 import functools
 import numbers
 import operator
-from typing import IO, Iterable, Literal, Mapping, Sequence, TypeVar
+from typing import Iterable, Literal, Mapping, Sequence, TypeVar
 
 import dateutil.parser
 import pandas as pd
 
 import ibis.common.exceptions as com
-import ibis.expr.analysis as _L
 import ibis.expr.builders as bl
 import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
@@ -3950,1047 +3949,8 @@ _category_value_methods = {'label': category_label}
 
 _add_methods(ir.CategoryValue, _category_value_methods)
 
-
-# ---------------------------------------------------------------------
-# Table API
-
-_join_classes = {
-    'inner': ops.InnerJoin,
-    'left': ops.LeftJoin,
-    'any_inner': ops.AnyInnerJoin,
-    'any_left': ops.AnyLeftJoin,
-    'outer': ops.OuterJoin,
-    'right': ops.RightJoin,
-    'left_semi': ops.LeftSemiJoin,
-    'semi': ops.LeftSemiJoin,
-    'anti': ops.LeftAntiJoin,
-    'cross': ops.CrossJoin,
-}
-
-
-def join(
-    left: ir.TableExpr,
-    right: ir.TableExpr,
-    predicates: str
-    | Sequence[
-        str
-        | tuple[str | ir.ColumnExpr, str | ir.ColumnExpr]
-        | ir.BooleanColumn
-    ] = (),
-    how: Literal[
-        'inner',
-        'left',
-        'outer',
-        'right',
-        'semi',
-        'anti',
-        'any_inner',
-        'any_left',
-        'left_semi',
-    ] = 'inner',
-    *,
-    suffixes: tuple[str, str] = ("_x", "_y"),
-) -> ir.TableExpr:
-    """Perform a join between two tables.
-
-    Parameters
-    ----------
-    left
-        Left table to join
-    right
-        Right table to join
-    predicates
-        Boolean or column names to join on
-    how
-        Join method
-    suffixes
-        Left and right suffixes that will be used to rename overlapping
-        columns.
-    """
-    klass = _join_classes[how.lower()]
-    if isinstance(predicates, ir.Expr):
-        predicates = _L.flatten_predicate(predicates)
-
-    expr = klass(left, right, predicates).to_expr()
-
-    # semi/anti join only give access to the left table's fields, so
-    # there's never overlap
-    if how in ("semi", "anti"):
-        return expr
-
-    return ops.relations._dedup_join_columns(
-        expr,
-        left=left,
-        right=right,
-        suffixes=suffixes,
-    )
-
-
-def asof_join(
-    left: ir.TableExpr,
-    right: ir.TableExpr,
-    predicates: str | ir.BooleanColumn | Sequence[str | ir.BooleanColumn] = (),
-    by: str | ir.ColumnExpr | Sequence[str | ir.ColumnExpr] = (),
-    tolerance: str | ir.IntervalScalar | None = None,
-    *,
-    suffixes: tuple[str, str] = ("_x", "_y"),
-) -> ir.TableExpr:
-    """Perform an "as-of" join between `left` and `right`.
-
-    Similar to a left join except that the match is done on nearest key rather
-    than equal keys.
-
-    Optionally, match keys with `by` before joining with `predicates`.
-
-    Parameters
-    ----------
-    left
-        Table expression
-    right
-        Table expression
-    predicates
-        Join expressions
-    by
-        column to group by before joining
-    tolerance
-        Amount of time to look behind when joining
-    suffixes
-        Left and right suffixes that will be used to rename overlapping
-        columns.
-
-    Returns
-    -------
-    TableExpr
-        Table expression
-    """
-    expr = ops.AsOfJoin(left, right, predicates, by, tolerance).to_expr()
-    return ops.relations._dedup_join_columns(
-        expr,
-        left=left,
-        right=right,
-        suffixes=suffixes,
-    )
-
-
-def cross_join(
-    left: ir.TableExpr,
-    right: ir.TableExpr,
-    *rest: ir.TableExpr,
-    suffixes: tuple[str, str] = ("_x", "_y"),
-) -> ir.TableExpr:
-    """Compute the cross join of a sequence of tables.
-
-    Parameters
-    ----------
-    left
-        Left table
-    right
-        Right table
-    rest
-        Additional tables to cross join
-    suffixes
-        Left and right suffixes that will be used to rename overlapping
-        columns.
-
-    Returns
-    -------
-    TableExpr
-        Cross join of `left`, `right` and `rest`
-
-    Examples
-    --------
-    >>> import ibis
-    >>> schemas = [(name, 'int64') for name in 'abcde']
-    >>> a, b, c, d, e = [
-    ...     ibis.table([(name, type)], name=name) for name, type in schemas
-    ... ]
-    >>> joined1 = ibis.cross_join(a, b, c, d, e)
-    >>> joined1  # doctest: +NORMALIZE_WHITESPACE
-    ref_0
-    UnboundTable[table]
-      name: a
-      schema:
-        a : int64
-    ref_1
-    UnboundTable[table]
-      name: b
-      schema:
-        b : int64
-    ref_2
-    UnboundTable[table]
-      name: c
-      schema:
-        c : int64
-    ref_3
-    UnboundTable[table]
-      name: d
-      schema:
-        d : int64
-    ref_4
-    UnboundTable[table]
-      name: e
-      schema:
-        e : int64
-    CrossJoin[table]
-      left:
-        Table: ref_0
-      right:
-        CrossJoin[table]
-          left:
-            CrossJoin[table]
-              left:
-                CrossJoin[table]
-                  left:
-                    Table: ref_1
-                  right:
-                    Table: ref_2
-              right:
-                Table: ref_3
-          right:
-            Table: ref_4
-    """
-    expr = ops.CrossJoin(
-        left,
-        functools.reduce(ir.TableExpr.cross_join, rest, right),
-        [],
-    ).to_expr()
-    return ops.relations._dedup_join_columns(
-        expr,
-        left=left,
-        right=right,
-        suffixes=suffixes,
-    )
-
-
-def _table_count(self: ir.TableExpr) -> ir.IntegerScalar:
-    """Compute the number of rows in `self`.
-
-    Parameters
-    ----------
-    self
-        Table expression
-
-    Returns
-    -------
-    IntegerScalar
-        Count of the number of rows in `self`
-    """
-    return ops.Count(self, None).to_expr().name('count')
-
-
-def _table_dropna(
-    self: ir.TableExpr,
-    subset: Sequence[str] | None = None,
-    how: Literal['any', 'all'] = 'any',
-) -> ir.TableExpr:
-    """Remove rows with null values from the table.
-
-    Parameters
-    ----------
-    subset
-        Columns names to consider when dropping nulls. By default all columns
-        are considered.
-    how
-        Determine whether a row is removed if there is at least one null
-        value in the row ('any'), or if all row values are null ('all').
-        Options are 'any' or 'all'. Default is 'any'.
-
-    Examples
-    --------
-    >>> import ibis
-    >>> t = ibis.table([('a', 'int64'), ('b', 'string')])
-    >>> t = t.dropna()  # Drop all rows where any values are null
-    >>> t = t.dropna(how='all')  # Only drop rows where all values are null
-    >>> t = t.dropna(subset=['a'], how='all')  # Only drop rows where all values in column 'a' are null  # noqa: E501
-
-    Returns
-    -------
-    TableExpr
-        Table expression
-    """
-    if subset is None:
-        subset = []
-    subset = util.promote_list(subset)
-    return ops.DropNa(self, how, subset).to_expr()
-
-
-def _table_fillna(
-    self: ir.TableExpr,
-    replacements: ir.ScalarExpr | Mapping[str, ir.ScalarExpr],
-) -> ir.TableExpr:
-    """Fill null values in a table expression.
-
-    Parameters
-    ----------
-    self
-        Table expression
-    replacements
-        Value with which to fill the nulls. If passed as a mapping, the keys
-        are column names that map to their replacement value. If passed
-        as a scalar, all columns are filled with that value.
-
-    Notes
-    -----
-    There is potential lack of type stability with the `fillna` API. For
-    example, different library versions may impact whether or not a given
-    backend promotes integer replacement values to floats.
-
-    Examples
-    --------
-    >>> import ibis
-    >>> t = ibis.table([('a', 'int64'), ('b', 'string')])
-    >>> t = t.fillna(0.0)  # Replace nulls in all columns with 0.0
-    >>> t.fillna({c: 0.0 for c, t in t.schema().items() if t == dt.float64})  # Replace all na values in all columns of a given type with the same value  # noqa: E501
-
-    Returns
-    -------
-    TableExpr
-        Table expression
-    """
-    if isinstance(replacements, collections.abc.Mapping):
-        columns = replacements.keys()
-        table_columns = self.schema().names
-        invalid = set(columns) - set(table_columns)
-        if invalid:
-            raise com.IbisTypeError(
-                f'value {list(invalid)} is not a field in {table_columns}.'
-            )
-    return ops.FillNa(self, replacements).to_expr()
-
-
-def _table_info(self: ir.TableExpr, buf: IO[str] | None = None) -> None:
-    """Show column names, types, and null counts.
-
-    Output to stdout by default.
-
-    Parameters
-    ----------
-    self
-        Table expression
-    buf
-        A writable buffer
-    """
-    metrics = [self.count().name('nrows')]
-    for col in self.columns:
-        metrics.append(self[col].count().name(col))
-
-    metrics = self.aggregate(metrics).execute().loc[0]
-
-    names = ['Column', '------'] + self.columns
-    types = ['Type', '----'] + [repr(x) for x in self.schema().types]
-    counts = ['Non-null #', '----------'] + [str(x) for x in metrics[1:]]
-    col_metrics = util.adjoin(2, names, types, counts)
-    result = f'Table rows: {metrics[0]}\n\n{col_metrics}'
-
-    print(result, file=buf)
-
-
-def _table_set_column(table: ir.TableExpr, name: str, expr: ir.ValueExpr):
-    """Replace an existing column with a new expression.
-
-    Parameters
-    ----------
-    table
-        Table expression
-    name
-        Column name to replace
-    expr
-        New data for column
-
-    Returns
-    -------
-    TableExpr
-        Table expression
-    """
-    expr = table._ensure_expr(expr)
-
-    if expr._name != name:
-        expr = expr.name(name)
-
-    if name not in table:
-        raise KeyError(f'{name} is not in the table')
-
-    # TODO: This assumes that projection is required; may be backend-dependent
-    proj_exprs = []
-    for key in table.columns:
-        if key == name:
-            proj_exprs.append(expr)
-        else:
-            proj_exprs.append(table[key])
-
-    return table.projection(proj_exprs)
-
-
-def _regular_join_method(
-    name: str,
-    how: Literal[
-        'inner',
-        'left',
-        'outer',
-        'right',
-        'semi',
-        'anti',
-        'any_inner',
-        'any_left',
-    ],
-):
-    def f(
-        self: ir.TableExpr,
-        other: ir.TableExpr,
-        predicates: str
-        | Sequence[
-            str
-            | tuple[str | ir.ColumnExpr, str | ir.ColumnExpr]
-            | ir.BooleanValue
-        ] = (),
-        suffixes: tuple[str, str] = ("_x", "_y"),
-    ) -> ir.TableExpr:
-        f"""Perform a{'n' * how.startswith(tuple("aeiou"))} {how} join between two tables.
-
-        Parameters
-        ----------
-        left
-            Left table to join
-        right
-            Right table to join
-        predicates
-            Boolean or column names to join on
-        suffixes
-            Left and right suffixes that will be used to rename overlapping
-            columns.
-
-        Returns
-        -------
-        TableExpr
-            Joined `left` and `right`
-        """  # noqa: E501
-        return self.join(other, predicates, how=how, suffixes=suffixes)
-
-    f.__name__ = name
-    return f
-
-
-def filter(
-    table: ir.TableExpr,
-    predicates: ir.BooleanValue | Sequence[ir.BooleanValue],
-) -> ir.TableExpr:
-    """Select rows from `table` based on `predicates`.
-
-    Parameters
-    ----------
-    table
-        Table expression
-    predicates
-        Boolean value expressions used to select rows in `table`.
-
-    Returns
-    -------
-    TableExpr
-        Filtered table expression
-    """
-    resolved_predicates = _resolve_predicates(table, predicates)
-    return _L.apply_filter(table, resolved_predicates)
-
-
-def _resolve_predicates(
-    table: ir.TableExpr, predicates
-) -> list[ir.BooleanValue]:
-    if isinstance(predicates, Expr):
-        predicates = _L.flatten_predicate(predicates)
-    predicates = util.promote_list(predicates)
-    predicates = [ir.relations.bind_expr(table, x) for x in predicates]
-    resolved_predicates = []
-    for pred in predicates:
-        if isinstance(pred, ir.AnalyticExpr):
-            pred = pred.to_filter()
-        resolved_predicates.append(pred)
-
-    return resolved_predicates
-
-
-def aggregate(
-    table: ir.TableExpr,
-    metrics: Sequence[ir.ScalarExpr] | None = None,
-    by: Sequence[ir.ValueExpr] | None = None,
-    having: Sequence[ir.BooleanValue] | None = None,
-    **kwargs: ir.ValueExpr,
-) -> ir.TableExpr:
-    """Aggregate a table with a given set of reductions grouping by `by`.
-
-    Parameters
-    ----------
-    table
-        Table to compute aggregates from
-    metrics
-        Aggregate expressions
-    by
-        Grouping expressions
-    having
-        Post-aggregation filters
-    kwargs
-        Named aggregate expressions
-
-    Returns
-    -------
-    TableExpr
-        An aggregate table expression
-    """
-    metrics = [] if metrics is None else util.promote_list(metrics)
-    metrics.extend(
-        table._ensure_expr(expr).name(name)
-        for name, expr in sorted(kwargs.items(), key=operator.itemgetter(0))
-    )
-
-    op = table.op().aggregate(
-        table,
-        metrics,
-        by=util.promote_list(by if by is not None else []),
-        having=util.promote_list(having if having is not None else []),
-    )
-    return op.to_expr()
-
-
-def _table_distinct(self: ir.TableExpr) -> ir.TableExpr:
-    """Compute the set of unique rows in the table."""
-    op = ops.Distinct(self)
-    return op.to_expr()
-
-
-def _table_limit(table: ir.TableExpr, n: int, offset: int = 0) -> ir.TableExpr:
-    """Select the first `n` rows at beginning of table starting at `offset`.
-
-    Parameters
-    ----------
-    table
-        Table expression
-    n
-        Number of rows to include
-    offset
-        Number of rows to skip first
-
-    Returns
-    -------
-    TableExpr
-        The first `n` rows of `table` starting at `offset`
-    """
-    op = ops.Limit(table, n, offset=offset)
-    return op.to_expr()
-
-
-def _head(table: ir.TableExpr, n: int = 5) -> ir.TableExpr:
-    """Select the first `n` rows of a table.
-
-    Notes
-    -----
-    The result set is not deterministic without a sort.
-
-    Parameters
-    ----------
-    table
-        Table expression
-    n
-        Number of rows to include, defaults to 5
-
-    Returns
-    -------
-    TableExpr
-        `table` limited to `n` rows
-
-    See Also
-    --------
-    ibis.expr.types.TableExpr.limit
-    """
-    return _table_limit(table, n=n)
-
-
-def _table_sort_by(
-    table: ir.TableExpr,
-    sort_exprs: str
-    | ir.ColumnExpr
-    | ir.SortKey
-    | tuple[str | ir.ColumnExpr, bool]
-    | Sequence[tuple[str | ir.ColumnExpr, bool]],
-) -> ir.TableExpr:
-    """Sort `table` by `sort_exprs`
-
-    Parameters
-    ----------
-    table
-        Table expression
-    sort_exprs
-        Sort specifications
-
-    Examples
-    --------
-    >>> import ibis
-    >>> t = ibis.table([('a', 'int64'), ('b', 'string')])
-    >>> ab_sorted = t.sort_by([('a', True), ('b', False)])
-
-    Returns
-    -------
-    TableExpr
-        Sorted `table`
-    """
-    return (
-        table.op()
-        .sort_by(
-            table,
-            [] if sort_exprs is None else util.promote_list(sort_exprs),
-        )
-        .to_expr()
-    )
-
-
-def _table_union(
-    left: ir.TableExpr,
-    right: ir.TableExpr,
-    distinct: bool = False,
-) -> ir.TableExpr:
-    """Compute the set union of two table expressions.
-
-    The input tables must have identical schemas.
-
-    Parameters
-    ----------
-    left
-        Table expression
-    right
-        Table expression
-    distinct
-        Only union distinct rows not occurring in the calling table (this
-        can be very expensive, be careful)
-
-    Returns
-    -------
-    TableExpr
-        Union of `left` and `right`
-    """
-    return ops.Union(left, right, distinct=distinct).to_expr()
-
-
-def _table_intersect(left: ir.TableExpr, right: ir.TableExpr) -> ir.TableExpr:
-    """Compute the set intersection of two table expressions.
-
-    The input tables must have identical schemas.
-
-    Parameters
-    ----------
-    left
-        Table expression
-    right
-        Table expression
-
-    Returns
-    -------
-    TableExpr
-        The rows common amongst `left` and `right`.
-    """
-    return ops.Intersection(left, right).to_expr()
-
-
-def _table_difference(left: TableExpr, right: TableExpr) -> ir.TableExpr:
-    """Compute the set difference of two table expressions.
-
-    The input tables must have identical schemas.
-
-    Parameters
-    ----------
-    left
-        Table expression
-    right
-        Table expression
-
-    Returns
-    -------
-    TableExpr
-        The rows present in `left` that are not present in `right`.
-    """
-    return ops.Difference(left, right).to_expr()
-
-
-def _table_to_array(self: ir.TableExpr) -> ir.ColumnExpr:
-    """View a single column table as an array.
-
-    Parameters
-    ----------
-    self
-        Table expression
-
-    Returns
-    -------
-    ValueExpr
-        A single column view of a table
-    """
-
-    schema = self.schema()
-    if len(schema) != 1:
-        raise com.ExpressionError(
-            'Table must have exactly one column when viewed as array'
-        )
-
-    return ops.TableArrayView(self).to_expr()
-
-
-def _safe_get_name(expr):
-    try:
-        return expr.get_name()
-    except com.ExpressionError:
-        return None
-
-
-def mutate(
-    table: ir.TableExpr,
-    exprs: Sequence[ir.Expr] | None = None,
-    **mutations: ir.ValueExpr,
-) -> ir.TableExpr:
-    """Add columns to a table expression.
-
-    Parameters
-    ----------
-    table
-        Table expression to add columns to
-    exprs
-        List of named expressions to add as columns
-    mutations
-        Named expressions using keyword arguments
-
-    Returns
-    -------
-    TableExpr
-        Table expression with additional columns
-
-    Examples
-    --------
-    Using keywords arguments to name the new columns
-
-    >>> import ibis
-    >>> table = ibis.table([('foo', 'double'), ('bar', 'double')], name='t')
-    >>> expr = table.mutate(qux=table.foo + table.bar, baz=5)
-    >>> expr  # doctest: +NORMALIZE_WHITESPACE
-    ref_0
-    UnboundTable[table]
-      name: t
-      schema:
-        foo : float64
-        bar : float64
-    <BLANKLINE>
-    Selection[table]
-      table:
-        Table: ref_0
-      selections:
-        Table: ref_0
-        baz = Literal[int8]
-          5
-        qux = Add[float64*]
-          left:
-            foo = Column[float64*] 'foo' from table
-              ref_0
-          right:
-            bar = Column[float64*] 'bar' from table
-              ref_0
-
-    Using the :meth:`ibis.expr.types.Expr.name` method to name the new columns
-
-    >>> new_columns = [ibis.literal(5).name('baz',),
-    ...                (table.foo + table.bar).name('qux')]
-    >>> expr2 = table.mutate(new_columns)
-    >>> expr.equals(expr2)
-    True
-
-    """
-    exprs = [] if exprs is None else util.promote_list(exprs)
-    for name, expr in sorted(mutations.items(), key=operator.itemgetter(0)):
-        if util.is_function(expr):
-            value = expr(table)
-        else:
-            value = rlz.any(expr)
-        exprs.append(value.name(name))
-
-    mutation_exprs = _L.get_mutation_exprs(exprs, table)
-    return table.projection(mutation_exprs)
-
-
-def projection(
-    table: ir.TableExpr,
-    exprs: ir.ValueExpr | str | Sequence[ir.ValueExpr | str],
-) -> ir.TableExpr:
-    """Compute a new table expression using `exprs`.
-
-    Parameters
-    ----------
-    exprs
-        Column expression, string, or list of column expressions and strings.
-
-    Returns
-    -------
-    TableExpr
-        Table expression
-
-    Notes
-    -----
-    Passing an aggregate function to this method will broadcast the aggregate's
-    value over the number of rows in the table and automatically constructs
-    a window function expression. See the examples section for more details.
-
-    Examples
-    --------
-    Simple projection
-
-    >>> import ibis
-    >>> fields = [('a', 'int64'), ('b', 'double')]
-    >>> t = ibis.table(fields, name='t')
-    >>> proj = t.projection([t.a, (t.b + 1).name('b_plus_1')])
-    >>> proj  # doctest: +NORMALIZE_WHITESPACE
-    ref_0
-    UnboundTable[table]
-      name: t
-      schema:
-        a : int64
-        b : float64
-    <BLANKLINE>
-    Selection[table]
-      table:
-        Table: ref_0
-      selections:
-        a = Column[int64*] 'a' from table
-          ref_0
-        b_plus_1 = Add[float64*]
-          left:
-            b = Column[float64*] 'b' from table
-              ref_0
-          right:
-            Literal[int8]
-              1
-    >>> proj2 = t[t.a, (t.b + 1).name('b_plus_1')]
-    >>> proj.equals(proj2)
-    True
-
-    Aggregate projection
-
-    >>> agg_proj = t[t.a.sum().name('sum_a'), t.b.mean().name('mean_b')]
-    >>> agg_proj  # doctest: +NORMALIZE_WHITESPACE, +ELLIPSIS
-    ref_0
-    UnboundTable[table]
-      name: t
-      schema:
-        a : int64
-        b : float64
-    <BLANKLINE>
-    Selection[table]
-      table:
-        Table: ref_0
-      selections:
-        sum_a = WindowOp[int64*]
-          sum_a = Sum[int64]
-            a = Column[int64*] 'a' from table
-              ref_0
-            where:
-              None
-          <ibis.expr.window.Window object at 0x...>
-        mean_b = WindowOp[float64*]
-          mean_b = Mean[float64]
-            b = Column[float64*] 'b' from table
-              ref_0
-            where:
-              None
-          <ibis.expr.window.Window object at 0x...>
-
-    Note the ``<ibis.expr.window.Window>`` objects here, their existence means
-    that the result of the aggregation will be broadcast across the number of
-    rows in the input column. The purpose of this expression rewrite is to make
-    it easy to write column/scalar-aggregate operations like
-
-    .. code-block:: python
-
-       t[(t.a - t.a.mean()).name('demeaned_a')]
-    """
-    import ibis.expr.analysis as L
-
-    if isinstance(exprs, (Expr, str)):
-        exprs = [exprs]
-
-    projector = L.Projector(table, exprs)
-    op = projector.get_result()
-    return op.to_expr()
-
-
-def _table_relabel(
-    table: ir.TableExpr, substitutions: Mapping[str, str]
-) -> ir.TableExpr:
-    """Change table column names, otherwise leaving table unaltered.
-
-    Parameters
-    ----------
-    table
-        Table expression
-    substitutions
-        Name mapping
-
-    Returns
-    -------
-    TableExpr
-        A relabeled table expression
-    """
-    observed = set()
-
-    exprs = []
-    for c in table.columns:
-        expr = table[c]
-        if c in substitutions:
-            expr = expr.name(substitutions[c])
-            observed.add(c)
-        exprs.append(expr)
-
-    for c in substitutions:
-        if c not in observed:
-            raise KeyError(f'{c!r} is not an existing column')
-
-    return table.projection(exprs)
-
-
-def _table_view(self) -> ir.TableExpr:
-    """Create a new table expression distinct from the current one..
-
-    Parameters
-    ----------
-    self
-        Table expression
-
-    Returns
-    -------
-    TableExpr
-        Table expression
-
-    Notes
-    -----
-    For doing any self-referencing operations, like a self-join, you will use
-    this operation to create a reference to the current table expression.
-    """
-    new_view = ops.SelfReference(self)
-    return new_view.to_expr()
-
-
-def _table_drop(
-    self: ir.TableExpr, fields: str | Sequence[str]
-) -> ir.TableExpr:
-    """Remove fields from a table.
-
-    Parameters
-    ----------
-    self
-        Table expression
-    fields
-        Fields to drop
-
-    Returns
-    -------
-    TableExpr
-        Expression without `fields`
-    """
-
-    if not fields:
-        # no-op if nothing to be dropped
-        return self
-
-    if isinstance(fields, str):
-        #  We want to drop just one attribute.
-        fields = [fields]
-
-    schema = self.schema()
-    field_set = frozenset(fields)
-    missing_fields = field_set.difference(schema)
-
-    if missing_fields:
-        raise KeyError(f'Fields not in table: {missing_fields!s}')
-
-    return self[[field for field in schema if field not in field_set]]
-
-
-def _rowid(self) -> ir.IntegerValue:
-    """An autonumbering expression representing the row number of the results.
-
-    It can be 0 or 1 indexed depending on the backend. Check the backend
-    documentation for specifics.
-
-    Notes
-    -----
-    This function is different from the window function `row_number`
-    (even if they are conceptually the same), and different from `rowid` in
-    backends where it represents the physical location
-    (e.g. Oracle or PostgreSQL's ctid).
-
-    Returns
-    -------
-    IntegerColumn
-        An integer column
-
-    Examples
-    --------
-    >>> my_table[my_table.rowid(), my_table.name].execute()
-    1|Ibis
-    2|pandas
-    3|Dask
-    """
-    return ops.RowID().to_expr()
-
-
-_table_methods = {
-    'aggregate': aggregate,
-    'count': _table_count,
-    'distinct': _table_distinct,
-    'drop': _table_drop,
-    'dropna': _table_dropna,
-    'fillna': _table_fillna,
-    'info': _table_info,
-    'limit': _table_limit,
-    'head': _head,
-    'set_column': _table_set_column,
-    'filter': filter,
-    'mutate': mutate,
-    'projection': projection,
-    'select': projection,
-    'relabel': _table_relabel,
-    'join': join,
-    'cross_join': cross_join,
-    'inner_join': _regular_join_method('inner_join', 'inner'),
-    'left_join': _regular_join_method('left_join', 'left'),
-    'right_join': _regular_join_method('right_join', 'right'),
-    'any_inner_join': _regular_join_method('any_inner_join', 'any_inner'),
-    'any_left_join': _regular_join_method('any_left_join', 'any_left'),
-    'outer_join': _regular_join_method('outer_join', 'outer'),
-    'semi_join': _regular_join_method('semi_join', 'semi'),
-    'anti_join': _regular_join_method('anti_join', 'anti'),
-    'asof_join': asof_join,
-    'sort_by': _table_sort_by,
-    'to_array': _table_to_array,
-    'union': _table_union,
-    'intersect': _table_intersect,
-    'difference': _table_difference,
-    'view': _table_view,
-    'rowid': _rowid,
-}
-
-
-_add_methods(ir.TableExpr, _table_methods)
-
-
-def prevent_rewrite(expr, client=None):
-    """Prevent optimization from happening below `expr`.
-
-    Parameters
-    ----------
-    expr : ir.TableExpr
-        Any table expression whose optimization you want to prevent
-    client : ibis.backends.base.Client, optional, default None
-        A client to use to create the SQLQueryResult operation. This is useful
-        if you're compiling an expression that derives from an
-        :class:`~ibis.expr.operations.UnboundTable` operation.
-
-    Returns
-    -------
-    sql_query_result : ir.TableExpr
-    """
-    if client is None:
-        client = expr._find_backend()
-    query = client.compile(expr)
-    return ops.SQLQueryResult(query, expr.schema(), client).to_expr()
+prevent_rewrite = ir.TableExpr.prevent_rewrite
+aggregate = ir.TableExpr.aggregate
+cross_join = ir.TableExpr.cross_join
+join = ir.TableExpr.join
+asof_join = ir.TableExpr.asof_join
